Try loading primitive replacements for all methods of a class, also when defined in super class
Normally, we were only loading primitives explicitly specified in the SOM core library. However, to be more flexible with the implementation, it is ok for a VM to provide additional primitives, for instance for performance reasons. Or, as for SOM++, we want also to be able to load them for functional reasons. Since String is now a super class of Symbol, we need to change the equality semantics, which is easiest when adding a VM-specific primitive for Symbol.

Signed-off-by: Stefan Marr <git@stefan-marr.de>