Remove BigInteger class, widden tagged and normal integers to 64bit
Tagged integers use 63 bit value range. The other bit can be represented with a instance of VMInteger. Since VMBigInteger does not implement arbitrary precision, we just remove it. Should be restored to provide 'infinite' integer value ranges.

Note, this means, SOM++ should be compiled as 64-bit version when tagged integers are used.

Note2: with the change to the latest core-lib, the equality for integers is purely based on value.
Signed-off-by: Stefan Marr <git@stefan-marr.de>