Pass interpreter to Invoke (which replaces the () operator)  and Send
I find the operator() the way it is used totally unreadable, so, now it is replaced.
Primitives should be able to use `this` to get the signature of their invocation etc. Don't think we need to pass it again a second time as parameter.

Furthermore, we changed from passing the frame to Send to actually passing the whole interpreter. This gives maximal flexibility.

Signed-off-by: Stefan Marr <git@stefan-marr.de>