Modify collectionLimit algorithm
Instead of basing collection limit on the survivor size just leave it
based on the heap size. The current algorithm would allow heap size
to grow if the workload is constantly growing.  Also it penalizes small
workloads as they can only allocate a small amount of memory before a GC
is started

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>