#contains: and #indexOf: can’t be supported as primitives
This removes #contains: and #indexOf: because they require a message send of #= to work, which means we can’t support it in the current SOM++ design.
When sending the message, we yield back to the bytecode loop, which is not recursive, and thus, can’t come back to the primitive code.